Fix some issues with infoboxes.
This fixes several issues with the new infobox implementation. To start with,
we now had one DOM element per possible infobox, which was a lot of elements
when looking at a dashboard. During the conversion, I also lost the timeout
before the infobox was shown. These two things combined so that if you quickly
scrolled your mouse across a bunch of users, it would pop up a ton of
infoboxes.

All the views for any given type now share a single element, and coordinate a
bit so that if you switch which user you're looking at (or which bug), the
infobox will update immedatiely rather than waiting for a new timeout.

This also fixes a small styling issue that presented when users had their
profile set to private.

Testing done:
Looked at a dashboard and tested several of the weird cases that arose when
mousing over the "submitter" column. Set my profile to private and looked at
the infobox that resulted.

Reviewed at https://reviews.reviewboard.org/r/8175/