Fix the order of captions within the image review UI.
When the image review UI is showing a diff, the captions were added in the
wrong order. We were correctly showing the images on the left and right but the
captions would show "(revision 1)" on the right and "(revision 2)" on the left.

Testing done:
Loaded a diff of two file attachment revisions. Saw that the before/after was
correct for both images and captions.

Reviewed at https://reviews.reviewboard.org/r/8215/