Fix the Diff Size column for review requests lacking that information.
The Diff Size column made some assumptions that certain statistical data
on the diffs were present, which wasn't the case for review requests without
diffs.

This adds some bullet-proofing to this column so that we only fetch the
information if it's available, and harmlessly fall back if not.

Testing Done:
Viewed the dashboard and saw the diff size info come up for all review
requests containing diffs, but not those that are just file aettachmnets.
Didn't see any errors in the console.

Reviewed at https://reviews.reviewboard.org/r/8094/