Fix the curl examples for custom Accept headers in the API docs.
The API docs were showing curl examples using -A for an Accept header,
but this is actually the User Agent header. This fixes the doc output to
show a correct example.

Testing Done:
Rebuilt the docs and tested the example for the `text/x-patch` header
on the diffs API. Saw the correct output.

Reviewed at https://reviews.reviewboard.org/r/8103/