Fix bugs bugs with local sites.
The bugs field on a review request had several bugs when using local sites. The
first of these is that we were passing in the arguments to `local_site_reverse`
as a tuple, but in the case where there's a local site, that gets appended to a
list (which doesn't work). Changing this to be a list fixes this bug. Second,
the `bug_url` and `bug_infobox` views took a `local_site` argument but didn't
add the necessary decorators to translate the `local_site_name` parameter into
a real object (and do associated access checks).

Bugs.

Testing done:
Set the bugs field on a review request on a local site. Saw that I could then
re-load the page and see the bugs listed. Clicked on a bug link and saw that I
was correctly redirected. Hovered over a link and saw a request to the infobox
URL that was correctly handled.

Reviewed at https://reviews.reviewboard.org/r/8224/