Expliclty use FilteredSelectMultiple for WebHook form's repo field
The repositories field of the `WebHookTargetForm` now explicitly uses a
`FilteredSelectMultiple` widget. While `WebHookAdmin` specifies that
this field in `filter_horizontal`, if any non-admin page should use this
form they will want to use a `FilteredSelectMultiple` for this widget.

Testing Done:
Tested this with an upcoming patch in RBCommons.

Reviewed at https://reviews.reviewboard.org/r/8225/