Fix WebHook signatures to follow our documentation.
Our documentation on the way WebHook payloads are signed claimed to use SHA-1
hashes for HMAC, but in reality, we were using MD5. This is because we were
failing to pass in the hash method to the HMAC library.

Testing done:
Ran unit tests.

Fixes bug 4412

Reviewed at https://reviews.reviewboard.org/r/8231/