Fix a regression in the display of the CodeMirror horizontal scrollbar.
The recent update to use the appropriate CodeMirror CSS file for our
version of CodeMirror forced the display of horizontal scrollbars. A
comment in their CSS says that this is meant to stay on, so presumably
some other internal machinery should be hiding the scrollbar in our
case, but that's not happening.

Since this CSS rule is new (from our perspective), and the styling we
had before didn't break anything, we're going to just force it off
again, like we did before.

Testing Done:
Tested a few of the CodeMirror editors throughout the product. They
behaved as they always have, and I didn't see the scrollbars.

Reviewed at https://reviews.reviewboard.org/r/8233/