Show a text cursor anywhere on an editable text area element.
When showing a text field in a non-editing state, we were showing a text
cursor, but only for the area of the field containing content. Any area
outside of that was getting the standard arrow pointer.

We now show a text cursor for the entire element, making it feel like a
real text field.

Testing Done:
Created a new review request and placed my cursor over an empty field.
Saw a text cursor instead of an arrow cursor.

Reviewed at https://reviews.reviewboard.org/r/8234/