Fix missing data for file attachment thumbnails for change descriptions.
If a file attachment thumbnail needed to make use of the User currently
browsing the page, it could break when rendering change descriptions, as
the request variable wasn't being provided to the template context.
The implementation would only see an empty string. This was causing
problems with Power Pack's PDF thumbnail support.

We now simply ensure that the request is passed in as part of the
context.

Testing Done:
Tested against a PDF thumbnail on a change description. I didn't see any
errors.

Had a customer test this as well, and he verified the errors went away on
his end.

Reviewed at https://reviews.reviewboard.org/r/8232/