Fix javascript serialization for old file attachment models without history.
In most places, old models without history work just fine, but in the case of
displaying thumbnails inside of change descriptions, it broke by spitting out
`attachmentHistoryID: None` into the javascript. This change conditionalizes
that serialization, such that it will just get the default from the model
attribute definition.

This was initially reviewed for RBCommons, and this change is just a crossport
to Review Board proper.

Testing done:
Added a file, published, removed it, published, and then removed its history
from the database. Loaded the review request page and saw that there were no
javascript errors.