Fix CSS linear gradient which resulted in strangely rendered buttons.
The move to less-plugin-autoprefix in 53bc220 exposed several
`linear-gradient()` calls that were being formulated incorrectly, resulting in
strangely rendered buttons in Firefox and IE.  The <side-or-corner> parameter
should be expressed in "**to** <side-or-corner>" syntax rather than the old
"<side>" syntax without the "to".

Changed all usages to "to" syntax.  For the `vertical-three-colors` case added
an explicit "to bottom", even though this was already the implicit default.

Also fixed an unrelated typo in defs.css to change "rgpa" to "rgba".

Testing Done:
Manually tested with Firefox 46.0.1, IE 11.0.9600, and Chrome 51.0.2704.
Previously rendering was fine in Chrome, but strange in Firefox and IE.  Now
buttons look fine in all three browsers.

Reviewed at https://reviews.reviewboard.org/r/8239/