Fix running the Review Board test suite using `manage.py test`.
The recent work to add modern Pipeline support needed to make use of the
flags we had for determining if we're running unit tests, and therefore
needed to move that state setting out of the test runner class and into
the actual test script being executed. The necessary change was made to
tests/runtests.py, but manage.py never got the fix.

We now set the RB_RUNNING_TESTS environment variable in manage.py if we
know we're running the test suite, enabling this script to be used
reliably once again.

Testing Done:
Unit tests run with manage.py.

Reviewed at https://reviews.reviewboard.org/r/8171/