Only show avatar settings form when avatars are enabled site-wide
The avatar settings form now only appears in the profile page when the
"Enable avatar services" settings is set in the admin panel. Otherwise,
it will not be available.

Testing Done:
Enabled and disabled avatars. The form was correctly shown or not
shown in each case.

Reviewed at https://reviews.reviewboard.org/r/8221/