Fix call to local_site_reverse in bugs column
The `BugsColumn` was calling `local_site_reverse` with `args` as a tuple
instead of a list, which would cause an exception when the repository
had a bug tracker configured. This has been corrected by updating the
calling point to use a list instead of a tuple and inside the
`local_site_reverse` to cast `args` to a list before concatenation.

The parameters of `local_site_reverse` have also been documented.

Testing Done:
- Ran unit tests.
- Viewed the dashboard with a review request filed against a repository
  with a bug tracker. Without this fix applied, the exception occured;
  with it applied, it did not.

Reviewed at https://reviews.reviewboard.org/r/8244/