Check if avatars are enabled at runtime instead of startup
Previously, we would add the `AvatarSettingsForm` to the `ProfilePage`
if, upon startup, the avatar services were enabled for the server.
However, this prevented creating a new database with `manage.py syncdb`.
We now use the `is_visible` method on the `ProfileForm` to determine
whether or not it should be shown.

Testing Done:
- Verified the avatar form is only shown when avatars are enabled.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8252/