Add visual feedback when clicking "Show content" for deleted files.
It's usually pretty quick to load the content for deleted files, but sometimes
things can be laggy. This change adds a spinner that's shown in the interval
between clicking "Show content" and the actual loading of said content.

Testing done:
Clicked a link. Saw a spinner.

Reviewed at https://reviews.reviewboard.org/r/8262/