Status API part 2: Small refactor of reviewtags.
A couple template tags in the reviewtags library (reply_section and reply_list)
took an entry as a parameter, but the only part of the entry that they cared
about was the review. This change refactors that to pass in the review instead.

I've also changed the review box template to directly reference the review as a
variable rather than repeatedly accessing entry.review.

Testing done:
- Verified that reviews and replies to reviews still displayed correctly.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8299/