Fix some JS issues on the dvcs branch
The dvcs branch had a few typos that slipped through review; this patch
addresses those.

In addition, two bugs have been fixed. The first is that the expanded
commit description would show if the description contained a trailing
newline. Now, we trim the description and summary before comparing them
to determine if we should show the extended description.

The second bug is that the diff viewer would attempt to view the commit
range of an interdiff, which is invalid. Now, we only inspect the commit
range when interdiffs are not selected.

Testing Done:
The above issues no longer occur.

Reviewed at https://reviews.reviewboard.org/r/8282/