Fix sending diff commit information from context API
The context API was only sending information about the commits when the
`base-commit-id` and `tip-commit-id` arguments were both present.
However, this is the opposite behaviour of what we desire. We do not
need information about commits when we already have it. Instead, we
require it when we know nothing about the commits.

Testing Done:
Previously, the diff commit index would not update when switching
revisions. It now switches revisions correctly.

Reviewed at https://reviews.reviewboard.org/r/8281/