Restore Jasmine 1.3.x behavior of hiding filtered-out suites.
Jasmine 1.4 changed the behavior for the ?spec= parameter, opting to
continue showing all suites instead of only those referenced in the
spec. This is problematic for very large test suites, as it can be hard
to locate the tests you're working with.

To get back the old behavior, this change introduces a new module,
jasmine.hide-filtered, which goes through and hides any suites that
don't contain at least one enabled test somewhere in its tree. This is
the new default behavior when this module is added, but that can be
toggled through a new option added to Jasmine's Options panel.

Testing Done:
Ran tests with the defaults, and saw that when a spec was specified,
only the relevant suites were shown. Saw that the checkbox for the
option was also checked.

Unchecked the checkbox and saw the page reload, re-running all tests,
and showing all suites, including those disabled. Saw that the checkbox
was still unchecked.

Re-toggled the checkbox, and saw the new default behavior and states
again.

Reviewed at https://reviews.reviewboard.org/r/8273/