Refactor review request header into a common template.
Every review request (and review-request like box such as the file review UI)
has a header that includes some icon-based actions, a bunch of text-based
actions, and some tabs. The code for these was spread out over every view and
template that wanted to include them, and the functionality was slightly
different for every implementation.

This change centralizes all of that into a single template which renders all of
the actions (including the "archive" menu, which was previously only available
on the main review request page). The tabs are handled somewhat specially by
passing in a list of tab data into the template context. The logic for the
"current tab" is done by comparing the request path to the tab's URL.

Testing done:
Checked the header display and functionality on the main review request page,
the diff viewer page (with different diff revisions), and the image review UI.

Reviewed at https://reviews.reviewboard.org/r/8307/