Clean up icon-based actions in review request header.
The CSS that handled the icon-based actions on the left-hand side of the header
(star and archive/mute) was pretty ugly. We had a mix of padding/margin
specifications, especially working around a bad value in icons.less for the
position of the archive icon. We also had rules spread around in various
places, including a couple that were completely obsolete.

This change includes the first introduction of review-request.less. My eventual
goal is to move everything corresponding to the review request box into there,
leaving reviews.less for the actual review/change description boxes.

Testing done:
Verified the actions' appearance in Chrome, Firefox, and Safari.

Reviewed at https://reviews.reviewboard.org/r/8308/