Port various review request detail JS files to ES6.
This change ports over several of the pieces of the review request detail page,
including the page view itself, and the review and changedesc box views. This
is in preparation for some refactoring to make the creation and plumbing of all
of these boxes much simpler.

Testing done:
- Smoke tested a page with multiple review and changedesc boxes.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/8316/