Refactor away ReviewBoxListView.
This change makes it so that instead of `ReviewBoxListView` inspecting the DOM
and creating models and views to go along with it, those models and views are
created directly with the appropriate data. After this, `ReviewBoxListView` was
pretty useless, and because `ReviewRequestPageView` already had its fingers in
the pie, I've just moved all its functionality into there.

Testing done:
- Ran js-tests.
- Smoke tested the review request detail page with multiple reviews and change
  description boxes.

Reviewed at https://reviews.reviewboard.org/r/8318/