Add codebase documentation for the core SCMTools classes.
People in the past have been confused as to what they need to override
or implement on SCMTool. A while back, I wrote documentation for most of
this, but never got it into a committable form. This finishes that work,
and brings it up to our current documentation standards.

All classes (SCMTool, Commit, Branch, etc.) and special Revisions in
reviewboard.scmtools.core are now documented, containing the
expectations for the consumers and subclasses, lists of expected
exceptions, and more.

Testing Done:
Generated the docs and went through them, looking for broken links or
other obvious errors.

Reviewed at https://reviews.reviewboard.org/r/8314/