Automatically install p4python for development installs.
Perforce support for Python used to be painful to install, but these
days it's a standard upstream package maintained by the Perforce team.
In order to improve test coverage on development installs, we're now
including this package in dev-requirements.txt, ensuring it'll be
installed by default when using prepare-dev.py.

This is a trivial change.