Status API part 3: Refactor entries on review detail page.
The review detail page is huge and ugly and multi-tentacled. This change
attempts to clean things up just a little bit, by creating some classes for the
various types of "entries" (the review and change boxes).

This is primarily code motion, but I did discover an obscure bug while I was
doing this related to default expansion for draft reply comments (the `entry`
variable was never assigned in that case, leaving it to be whatever it was
before).

There's obviously a ton more that could be done here, but I want to take this
one step at a time.

Testing done:
Verified the appearance of the review detail page with multiple reviews,
replies, and change descriptions. Everything works as it did before.

Reviewed at https://reviews.reviewboard.org/r/8339/