Handle HostingServiceErrors in the the repo branch resource
Previously, if there was an error talking to the hosting service while
fetching the branches we would not handle the error and the user would
only see that a HTTP 500 error. We now present catch this error and
present it to the user.

Testing Done:
Faked an API error in `GitHub.get_branches` with and without this
patch applied. Without the patch applied a 500 error was shown. With
this patch, the error message was shown.

Reviewed at https://reviews.reviewboard.org/r/8340/