Add support for Splat as a bug tracker
The patch adds support for using Splat as a bug tracker for a
repository. As Splat is a SaaS, only the https://hellosplat.com instance
is supported.

Testing done:
Added Splat as a bug tracker and saw the correct information in the bug
infobox. The bug URL view redirected to the correct place.