Fix unit tests for the auth backend registry
The tests for the AuthBackendRegistry were incorrectly assuming only
that `StandardAuthBackend` would be registered, when in actuality all
backends listed in the set of `reviewboard.auth.backends` entrypoint
would be registered. Instead of comparing against
`{StandardAuthBackend}`, we compute the set of registered auth backends
before performing registration and unregistration operations on the
registry.

Testing done:
Ran unit tests.