Don't always show the vertical scrollbar on the user selector.
The new user selector was always showing space for a vertical scrollbar,
which looked bad on any OS that shows scrollbars inline. This was due to
an `overflow-y: scroll` rule.

This change makes the scrollbar only appear when necessary, using `auto`
instead. If there isn't enough room for the list, the scrollbar will
kick in, but otherwise it will be hidden.

Testing Done:
Looked at the user selector and saw that the contents of the list were
flush with the sides. No scrollbar space was being reserved.

Forced the height of the list to be too small for the contents, and saw
the scrollbar appear.

Reviewed at https://reviews.reviewboard.org/r/8355/