Split reviews/tests.py into multiple test files.
The reviews/tests.py file was getting large, and an upcoming change was
about to increase the size considerably. Along with this, it's become a
bit cluttered and disorganized in places. This all makes it harder for
new contributors (and us old-timers) to contribute new unit tests in a
consistent way.

This change is the first in a series that will clean up these unit
tests. This change simply splits the tests.py file into a handful of
individual files, based on the area of responsibility.

The code has simply moved in this change, without modern code standards
being applied. This is intentional, to help with backporting to other
branches. Future changes will clean up the test suite.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/8359/