Add an admin database page for viewing integration configurations.
This adds a page in the admin UI's Database section for viewing any
configured integration configurations. This is of course not the
preferred way of managing integration configurations, but is useful for
debugging, and for cleaning up entries from integrations that no longer
exist.

Testing Done:
Viewed my integrations in the database viewer.

Reviewed at https://reviews.reviewboard.org/r/8371/