Remove field-based filtering of integration configurations.
The `Integration.get_configs()` in Djblets allows for filtering of
configurations based on fields on the configuration model. Review Board
adopted this as well, but we don't actually want to expose this.
Instead, callers should be operating based on the contents of the
configuration's settings, which must be done manually anyway.

This cleans up the documentation for this method, and removes the custom
filtering.

Testing Done:
Tested that the Slack integration still worked.

Reviewed at https://reviews.reviewboard.org/r/8370/