Fix some issues with screenshot commenting.
This fixes a couple bugs that have snuck in when commenting on legacy
screenshots. The first one was that the screenshot comment model doesn't define
`canUpdateBounds`, which was causing an error trying to call it. While fixing
that I fixed up some documentation on the matching version for image file
attachments. The second bug was that the screenshot model doesn't have a
`filename` attribute, which was causing a similar error when trying to decide
the default zoom level to use.

Testing done:
Was able to comment on a screenshot without errors. Verified, to my surprise,
that the new zoom feature worked without issues.

Reviewed at https://reviews.reviewboard.org/r/8385/