Status Updates part 8: User Interface.
This change adds the UI for status updates in the review request page. Status
updates are shown in two places:

- If a status update is associated with a `ChangeDescription`, it will be shown
  in the entry for that change description.
- Otherwise, it's assumed that the status update is for the initial diff (or any
  other attachments, etc. that were present when the review request was first
  published). These are shown in a new entry which always appears immediately
  below the review request box.

The one thing that's missing from this is the draft banners when replying to
any comments in the reviews associated with status updates. This is because
multiple reviews can be aggregated together into one entry, and we don't yet
have the ability to deal with that. Our proposed redesign for the draft banners
will solve this (since there will always be one banner at the top of the page).

If the `reviews.status_updates` feature is not enabled (which is currently the
default), any reviews associated with status updates will be displayed just
like all the other reviews, in their own entries.

Testing done:
- Created a variety of status updates and checked that they rendered correctly.
- Tested the default expand/collapse state for the new initial status updates
  entry (expanded only when there are no change descriptions or draft replies).
- Disabled the feature in my settings and saw that all the reviews were now
  shown as regular review entries.

Reviewed at https://reviews.reviewboard.org/r/8409/