Remove a trailing comma in the JavaScript for the dashboard page.
The dashboard sets up some JavaScript to govern the functionality of the
page, but for non-Local Sites it ends up with a trailing comma, which
could break some browsers.

This change switches the order of attributes passed to RB.Dashboard,
preventing the trailing comma issue from occurring.

Testing Done:
Checked the Dashboard page and source for the global and local sites, and
saw that the pages both worked and didn't have trailing commas.

Reviewed at https://reviews.reviewboard.org/r/8381/